+++ developers-api/README.md
@@ -52,3 +52,10 @@ If you want to learn more about building native executables, please consult http
 REST is easy peasy with this Hello World RESTEasy resource.
 
 [Related guide section...](https://quarkus.io/guides/getting-started#the-jax-rs-resources)
+
+
+### Running the H2 DB console
+
+The H2 DB console is available at ```http://localhost:8080/h2```.  
+
+Log in using jdbc:h2:mem:test with an empty username and password.  A Developers tasble is present with preloaded data.
