+++ developers-api/src/main/resources/META-INF/web.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE web-app
+        PUBLIC "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN"
+        "http://java.sun.com/dtd/web-app_2_3.dtd">
+
+<web-app>
+    <display-name>My Web Application</display-name>
+
+    <servlet>
+        <servlet-name>h2-console</servlet-name>
+        <servlet-class>org.h2.server.web.WebServlet</servlet-class>
+    </servlet>
+
+    <servlet-mapping>
+        <servlet-name>h2-console</servlet-name>
+        <url-pattern>/h2/*</url-pattern>
+    </servlet-mapping>
+
+</web-app>
